Contacts Activity mostly implemented
Contacts model, and main activity created.

The main activity uses a top menu with two buttons below that for adding a new group and a new contact.

There is a recycler view below that which presents all a users contacts in this system.

The new group button brings up a new activity with a top menu and two recycler views, top one horizontal for selected contatcs, and the bottom one vertical again showing all contacts.

The new contacts brings up another activity with a form for adding a new user, first name, last name, username, and email.