All connected
Everything is connected and should work.  We will know better when real data is attached.

Created a new adapter to work with the new groups contacts. could have used the same one, but that would mean a lot of findangling, easier to just basically copy the contact adapter and modify the interface and onClick.