Firebase Api mostly working
Some recyclerview duplication on contacts page, only coming from chatfragment.

and

same for messages only when the send button is pressed.

all duplications disapear wehen activity is opened from another venu. or not from a button press for the message send.

The chat window was working for everything but now only for newGroup.