Notification completed
Notifications are working as expected. when a use clicks on the notification it opens the message activity and loads that chat.

Had to create a workaround for notifications as the first time the call back is initialized a notification happens. Fixed this with a boolean and count.

tried using a service but in the end figured out we didn't really need a custom service to make it work.

User gets notification when program is in the foreground or background.